refactor: 텍스트 버튼으로 이름 변경
